Generate induction schemes #159
