+++ example/Proofs/InductionSchemes.hs
